Updated Loading state
